Week4 Day2/config.py
from dotenv import load_dotenv
import os

load_dotenv()  # Loads environment variables from the .env file

class Config:
    # MongoDB Configuration
    MONGO_URI = os.getenv('MONGO_URI', 'mongodb://localhost:27017/idea_vault')
    
    # JWT Configuration
    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY', 'your-secret-key')
    
    # Other Configs (if any)
    DEBUG = os.getenv('DEBUG', True)
    SECRET_KEY = os.getenv('SECRET_KEY', 'your-flask-secret-key')
