Week4 Day2/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
flask-cors==4.0.0       
Flask-RESTful==0.3.10  
pymongo==4.6.3          